requirements.txt
flask==2.3.2
flask-cors==3.0.10
PyPDF2==3.0.1
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
ollama==0.1.8
langchain-huggingface==0.0.2
python-dotenv==1.0.0
pytest==7.4.0
black==23.10.0